Rework this repository
- Functions in Account class have no more RequestResult returns. This class has been removed.
- Introducing AccountSession class. This will resolve login failed in the future (if not too much).
- Rework Login function using AccountSession class.
- Moved all functions in model to account_object.dart, account_session_object.dart, news_object.dart and utils_object.dart.
- Another class in model will moved to root.
- Optimize codes.
- Updated dependencies to latest version.
