+++ README.md
@@ -1,31 +1,44 @@
 # DutWrapper
 
-This library provides wrapper (for this repository - crawl data from a page) for access for some features in [DUT Student page - Da Nang University of Technology](http://sv.dut.udn.vn).
+An unofficial wrapper for easier to use at [sv.dut.udn.vn - Da Nang University of Technology student page](http://sv.dut.udn.vn).
 
-## Version
+# Version
 
-[![https://github.com/dutwrapper/dutwrapper-dart](https://img.shields.io/github/v/release/dutwrapper/dutwrapper-dart)](https://github.com/dutwrapper/dutwrapper-dart/releases)
+[![](https://img.shields.io/github/v/release/dutwrapper/dutwrapper-dart?label=release)](https://github.com/dutwrapper/dutwrapper-dart/releases)
 
-## Building requirements
+[![](https://img.shields.io/github/v/tag/dutwrapper/dutwrapper-dart?label=pre-release)](https://github.com/dutwrapper/dutwrapper-dart/releases)
 
+# Building requirements
 - Follow [this document](https://docs.flutter.dev/) for building or using this library.
 
-## License and credits
+# FAQ
 
-- [MIT](LICENSE) (click to view licenses)
+## Where can I found library changelog?
+If you want to:
+- View major changes: [Click here](CHANGELOG.md).
+- View entire source code changes: [Click here](https://github.com/dutwrapper/dutwrapper-dotnet/commits).
+  - You will need to change branch if you want to view changelog for stable/draft version.
 
-## FAQ
+## Branch in dutwrapper?
+- `stable`/`main`: Default branch and main release.
+- `draft`: Alpha branch. This branch is used for update my progress and it's very unstable. Use it at your own risk.
 
-### Branch in this repository?
-- `main`: Default branch and main release.
-- `draft`: Alpha branch. This branch is used for update my working and may not be tested. Use it at your own risk.
+## I received error about login while running AccountTest?
+- Did you mean this error: `dut_account environment variable not found. Please, add or modify this environment in format "username|password"`?
+- If so, you will need to add environment variable named `dut_account` with syntax `studentid|password`.
 
-### Wiki, or manual for how-to-use?
+## Wiki, or manual for how-to-use?
 - In a plan, please be patient.
 
-### Latest change log?
-- To view log for all versions, [click here](CHANGELOG.md)
+## I'm got issue with this library. Which place can I reproduce issue for you?
+- If you found a issue, you can report this via [issue tab](https://github.com/dutwrapper/dutwrapper-dotnet/issues) on this repository.
 
-## Copyright?
-- This project - dutwrapper - is not affiliated with [Da Nang University of Technology](http://sv.dut.udn.vn).
-- DUT, Da Nang University of Technology, web materials and web contents are trademarks and copyrights of [Da Nang University of Technology](http://sv.dut.udn.vn) school.
+# Credit and license?
+- License: [**MIT**](LICENSE)
+- DISCLAIMER:
+  - This project - dutwrapper - is not affiliated with [Da Nang University of Technology](http://sv.dut.udn.vn).
+  - DUT, Da Nang University of Technology, web materials and web contents are trademarks and copyrights of [Da Nang University of Technology](http://sv.dut.udn.vn) school.
+- Used third-party dependencies:
+  - [html](https://pub.dev/packages/html): Licensed under the [license](https://pub.dev/packages/html/license)
+  - [http](https://pub.dev/packages/http): Licensed under the [BSD-3-Clause license](https://pub.dev/packages/http/license)
+  - [collection](https://pub.dev/packages/collection): Licensed under the [BSD-3-Clause license](https://pub.dev/packages/collection/license)
