+++ test/news_test.dart
@@ -1,4 +1,8 @@
-import 'dart:developer';
+
+// ignore_for_file: avoid_print
+// This is already test file, we need to all log here
+
+import 'dart:convert';
 
 import 'package:dutwrapper/news.dart';
 import 'package:flutter_test/flutter_test.dart';
@@ -6,58 +10,56 @@ import 'package:flutter_test/flutter_test.dart';
 void main() {
   test('News Global', () async {
     for (int i = 1; i <= 5; i++) {
-      log('======= GET GLOBAL NEWS - PAGE $i ========');
+      print('======= GET GLOBAL NEWS - PAGE $i ========');
       final response = await News.getNewsGlobal(page: i);
 
       if (response.isNotEmpty) {
-        log('Subject list: ${response.length}');
+        print('Subject list: ${response.length}');
         for (var element in response) {
-          log('========================================');
-          log('Date: ${element.date}');
-          log('Title: ${element.title}');
-          log('Content String: ${element.contentString}');
-          log('Links:');
-          for (var link in element.links) {
-            log('========= Link ========');
-            log('Text: ${link.text}');
-            log('Position: ${link.position}');
-            log('Url: ${link.url}');
+          print('========================================');
+          print('Date: ${element.date}');
+          print('Title: ${element.title}');
+          print('Content: ${element.content}');
+          for (var link in element.resources) {
+            print('Link: ${link.position} - ${link.type} - ${link.text} - ${link.content}');
           }
         }
+
+        print(jsonEncode(response));
       } else {
-        log('Nothing in list!');
+        print('Nothing in list!');
       }
     }
   });
 
   test('News Subject', () async {
     for (int i = 1; i <= 5; i++) {
-      log('======= GET SUBJECT NEWS - PAGE $i =======');
+      print('======= GET SUBJECT NEWS - PAGE $i =======');
       final response = await News.getNewsSubject(page: i);
 
       if (response.isNotEmpty) {
-        log('Subject list: ${response.length}');
+        print('Subject list: ${response.length}');
         for (var element in response) {
-          log('========================================');
-          log('Date: ${element.date}');
-          log('Title: ${element.title}');
-          log('Content: ${element.contentString}');
-          log('Lecturer Gender: ${element.lecturerGender.toString()}');
-          log('Lecturer Name: ${element.lecturerName}');
-          log('Lesson Status: ${element.lessonStatus.toString()}');
-          log('Affected Date: ${element.affectedDate}');
-          log('Affected Lesson: ${element.affectedLessons.toString()}');
-          log('Affected Room: ${element.affectedRoom}');
-          log('Affected Class:');
-          for (var affectedClassItem in element.affectedClasses) {
-            log(affectedClassItem.subjectName);
-            for (var codeItem in affectedClassItem.codeList) {
-              log(codeItem.toStringTwoLastDigit());
-            }
+          print('========================================');
+          print('Date: ${element.date}');
+          print('Title: ${element.title}');
+          print('Content: ${element.content}');
+          for (var link in element.resources) {
+            print('Link: ${link.position} - ${link.type} - ${link.text} - ${link.content}');
           }
+          for (var affectedClassItem in element.affectedClasses) {
+            print("Class affected: ${affectedClassItem.subjectName} - ${affectedClassItem.codeList.map((p) => "${p.studentYearId}-${p.classId}").toList().join(", ")}");
+          }
+          print('Lecturer Gender: ${element.lecturerGender.toString()}');
+          print('Lecturer Name: ${element.lecturerName}');
+          print('Lesson Status: ${element.lessonStatus.toString()}');
+          print('Affected Date: ${element.affectedDate}');
+          print('Affected Lesson: ${element.affectedLessons.toString()}');
+          print('Affected Room: ${element.affectedRoom}');
         }
+        print(jsonEncode(response));
       } else {
-        log('Nothing in list!');
+        print('Nothing in list!');
       }
     }
   });
